feat: add running heatmap
